This change deals primarily with allowing developers to extend custom sections and controls with their own Backbone views.  Because we want third-party code to extend Backbone views rather than manipulating the DOM, it means that we need to render our views later to make sure that we're not rendering before those scripts have had a chance.
So, the rendering is taking place on `DOMContentLoaded`.  This means that scripts that are manipulating the DOM can be affected if they're running too early.  If developers need to manipulate the DOM for some reason, it should be done on `window.onload` or after.

This is not necessarily the final stage of these changes.  An alternative idea would be to render very late (hook into `admin_enqueue_scripts` with a priority of something like `95`).

Either way, it's best to extend the Backbone Views if at all possible rather than manipulating the DOM.