This is the last breaking JavaScript change, I promise. :)
This commit renames the `butterbean.views.add_*()` functions to `butterbean.views.register_*()` and the `butterbean.views.remove_*()` functions to `butterbean.views.unregister_*()`.

Yeah, it's a pedantic change, but it'll drive me crazy for years to come.  It's best to go ahead and make this consistent with the other `register_()` and `unregister_()` functions on the PHP side now before officially releasing.